image downsize and migrations renamed and remade
