+++ portal/static/portal/sass/partials/_banners.scss
@@ -86,7 +86,7 @@
 }
 
 .banner--picture--coding-club {
-  background-image: url("../img/jeswin-thomas-cropped.jpg")
+  background-image: url("../img/coding_club_hero.jpg")
 }
 
 .banner--picture--play {
