+++ portal/templates/portal/coding_club.html
@@ -78,7 +78,7 @@ function helpMe() {
         <div class="row d-flex">
             <div class="col-sm-6 d-flex flex-column">
                 <div class="flex-grow-1">
-                    <img title="Python coding club" alt="Python coding club" src="{% static 'portal/img/campaign-creators.png' %}">
+                    <img title="Python coding club" alt="Python coding club" src="{% static 'portal/img/coding_club_python_pack.png' %}">
                 </div>
             </div>
             <div class="col-sm-6 d-flex flex-column">
